+++ dev/test-studio/sanity.types.ts
@@ -12,6 +12,8 @@
  * ---------------------------------------------------------------------------------
  */
 
+export declare const internalGroqTypeReferenceTo: unique symbol
+
 // Source: schema.json
 export type Rendition = {
   publicuri?: string
@@ -266,6 +268,26 @@ export type RichDate = {
   offset?: number
 }
 
+export type I18nArrayCircularSchemaRepro = {
+  _id: string
+  _type: 'i18nArrayCircularSchemaRepro'
+  _createdAt: string
+  _updatedAt: string
+  _rev: string
+  title?: string
+  body?: I18nArrayCircularBodyContent
+}
+
+export type I18nArrayCircularBodyContent = Array<{
+  tabs?: Array<{
+    body?: I18nArrayCircularBodyContent
+    _type: 'tab'
+    _key: string
+  }>
+  _type: 'tabs'
+  _key: string
+}>
+
 export type I18nArrayPerformanceTest = {
   _id: string
   _type: 'i18nArrayPerformanceTest'
@@ -338,11 +360,13 @@ export type InternationalizedArrayText = Array<
 export type InternationalizedArrayTextValue = {
   _type: 'internationalizedArrayTextValue'
   value?: string
+  language: string
 }
 
 export type InternationalizedArrayStringValue = {
   _type: 'internationalizedArrayStringValue'
   value?: string
+  language: string
 }
 
 export type TranslationMetadata = {
@@ -371,6 +395,7 @@ export type LessonReference = {
 export type InternationalizedArrayReferenceValue = {
   _type: 'internationalizedArrayReferenceValue'
   value?: LessonReference
+  language: string
 }
 
 export type Lesson = {
@@ -643,6 +668,8 @@ export type AllSanitySchemaTypes =
   | Code
   | RichDateTest
   | RichDate
+  | I18nArrayCircularSchemaRepro
+  | I18nArrayCircularBodyContent
   | I18nArrayPerformanceTest
   | InternationalizedArrayString
   | InternationalizedPost
@@ -676,5 +703,3 @@ export type AllSanitySchemaTypes =
   | SanityAssetSourceData
   | SanityImageAsset
   | Geopoint
-
-export declare const internalGroqTypeReferenceTo: unique symbol