+++ plugins/@sanity/document-internationalization/src/actions/DeleteTranslationAction.test.tsx
@@ -232,6 +232,7 @@ describe('DeleteTranslationAction', () => {
         schemaTypes: ['article'],
         translations: [
           {
+            _key: 'random-key',
             [LANGUAGE_FIELD_NAME]: 'en',
             _type: 'internationalizedArrayReferenceValue',
             value: {_type: 'reference', _ref: 'doc-1'},
@@ -291,6 +292,7 @@ describe('DeleteTranslationAction', () => {
         translations: [
           {
             [LANGUAGE_FIELD_NAME]: 'en',
+            _key: 'random-key',
             _type: 'internationalizedArrayReferenceValue',
             value: {_type: 'reference', _ref: 'doc-1'},
           },