+++ plugins/sanity-plugin-internationalized-array/src/constants.ts
@@ -3,10 +3,27 @@ import type {PluginConfig} from './types'
 /**
  * The field name used to identify the language of an internationalized array item.
  *
- * In v4.x this was '_key', in v5+ this will be 'language'.
- * Having this as a constant makes the codebase easier to maintain.
+ * In v4.x this was '_key', in v5+ this is 'language'.
+ * ```ts
+ * {
+ *    "description": [
+ *      {
+ *        "_key": "kjjNvZHK8Y2QpTEf3K5jc",
+ *        "_type": "internationalizedArrayTextValue",
+ *        "language": "en"
+ *        "value": "This is the description in English"
+ *      },
+ *      {
+ *        "_key": "kjjNvZHK8Y2QpTEf3K5jc",
+ *        "_type": "internationalizedArrayTextValue",
+ *        "language": "es"
+ *        "value": "This is the description in Spanish"
+ *      },
+ *    ]
+ * }
+ * ```
  */
-export const LANGUAGE_FIELD_NAME = '_key' as const
+export const LANGUAGE_FIELD_NAME = 'language' as const
 
 export const MAX_COLUMNS = {
   codeOnly: 5,
@@ -24,4 +41,7 @@ export const CONFIG_DEFAULT: Required<PluginConfig> = {
   buttonAddAll: true,
   languageDisplay: 'codeOnly',
   includeForDocumentType: (documentType) => documentType !== 'translation.metadata',
+  languageFilter: {
+    documentTypes: [],
+  },
 }