+++ plugins/sanity-plugin-internationalized-array/src/index.ts
@@ -1,4 +1,5 @@
 export {clear} from './cache'
-export {internationalizedArray} from './plugin'
 export {LANGUAGE_FIELD_NAME} from './constants'
+export {internationalizedArray} from './plugin'
 export * from './types'
+export {internationalizedArrayLanguageFilter} from './utils/internationalizedArrayLanguageFilter'